Upgraded to scodec-bits 1.1.0

+++ src/main/scala/scalaz/stream/tcp.scala
@@ -171,8 +171,8 @@ object tcp {
            timeout: Option[Duration] = None,
            allowPeerClosed: Boolean = false): Process[Connection,Option[ByteVector]] =
     available(numBytes, timeout, allowPeerClosed) flatMap {
-      _ map { bs => if (bs.size == numBytes) Process.emit(Some(bs))
-                    else read(numBytes - bs.size, timeout, allowPeerClosed).map(_.map(bs ++ _))
+      _ map { bs => if (bs.size.toInt == numBytes) Process.emit(Some(bs))
+                    else read(numBytes - bs.size.toInt, timeout, allowPeerClosed).map(_.map(bs ++ _))
             } getOrElse (Process.emit(None))
     }
